changed scale factor

+++ scripts/demo.py
@@ -158,11 +158,11 @@ def scale_coords():
     top_left_corner = (min(x for x, y in draw_coords), min(y for x, y in draw_coords))
     boottom_right_corner = (max(x for x, y in draw_coords), max(y for x, y in draw_coords))
     center = ((top_left_corner[0] + boottom_right_corner[0]) // 2, (top_left_corner[1] + boottom_right_corner[1]) // 2)
-    scale_factor = min(WIDTH // (1 if boottom_right_corner[0] - top_left_corner[0] == 0 else boottom_right_corner[0] - top_left_corner[0]), 
-                       HEIGHT // (1 if boottom_right_corner[1] - top_left_corner[1] == 0 else boottom_right_corner[1] - top_left_corner[1]))
+    scale_factor = min(WIDTH // (boottom_right_corner[0] - top_left_corner[0] + 1), 
+                       HEIGHT // (boottom_right_corner[1] - top_left_corner[1] + 1)) / 1.5
     for x, y in draw_coords:
-        scaled_x = (x - center[0]) * scale_factor + WIDTH // 2
-        scaled_y = (y - center[1]) * scale_factor + HEIGHT // 2
+        scaled_x = int((x - center[0]) * scale_factor + WIDTH // 2)
+        scaled_y = int((y - center[1]) * scale_factor + HEIGHT // 2)
         scaled_coords.append((scaled_x, scaled_y))
 
 def apply_brush(grid, grid_x, grid_y):
